requirements.txt
Django==1.5.1
PIL==1.1.7
South==0.8.1
django-autoslug==1.7.1
django-crispy-forms==1.3.2
django-extensions==1.1.1
django-google-tools==1.0.0
django-registration==1.0
django-social-auth==0.7.25
django-tinymce==1.5.1
easy-thumbnails==1.3
httplib2==0.8
oauth2==1.5.211
psycopg2==2.5.1
py-bcrypt==0.4
python-openid==2.2.5
six==1.3.0
wsgiref==0.1.2
